requirements.txt
fastapi>=0.104.0
uvicorn>=0.23.2
python-dotenv>=1.0.0
langchain>=0.1.0,<0.4.0
langchain-openai>=0.0.5
langchain-groq>=0.2.4
groq>=0.4.1
langchain-community>=0.0.13
langchain-core>=0.3.33
qdrant-client>=1.7.0
tiktoken>=0.5.2
pydantic>=2.4.2,<3.0.0
supabase>=2.0.0
aiofiles>=23.2.1 